+++ packages/webcomponents/src/components/payouts-list/test/payouts-list-core.spec.tsx
@@ -43,7 +43,7 @@ describe('payouts-list-core', () => {
 
     // components.pop();
     const page = await newSpecPage({
-      components: [PayoutsListCore, PaginationMenu, TableFiltersMenu, PayoutsListFilters],
+      components: [PayoutsListCore, PaginationMenu],
       template: () => <payouts-list-core getPayouts={getPayouts} getSubAccounts={getSubAccounts} columns={defaultColumnsKeys} />,
     });
 
@@ -79,7 +79,7 @@ describe('payouts-list-core', () => {
     });
 
     const page = await newSpecPage({
-      components: [PayoutsListCore, PaginationMenu, TableFiltersMenu, PayoutsListFilters],
+      components: [PayoutsListCore, PaginationMenu],
       template: () => <payouts-list-core getPayouts={getPayouts} getSubAccounts={getSubAccounts} columns={defaultColumnsKeys} />,
     });